feat: secure server actions

+++ code-review-admin/src/actions/prompts.ts
@@ -1,6 +1,9 @@
 "use server"
 import type { Prompt } from "@/src/utils/api"
 import { commonHeaders } from "@/src/utils/api"
+import { auth } from "@/src/utils/auth"
+
+const unauthorized = "Unauthorized"
 
 export const getPrompts = async () => {
     return fetch(`${process.env.API_URL}/api/prompts`, {
@@ -8,7 +11,11 @@ export const getPrompts = async () => {
     }).then(res => res.json())
 }
 
-export const createPrompt = async (prompt: Prompt) => {
+export const createPrompt = async (
+    prompt: Prompt
+): Promise<{ prompt: Prompt | null; message: string }> => {
+    if (!await auth()) return { prompt: null, message: unauthorized }
+
     return fetch(`${process.env.API_URL}/api/prompts/${prompt.name}`, {
         method: "PUT",
         headers: commonHeaders,
@@ -16,7 +23,17 @@ export const createPrompt = async (prompt: Prompt) => {
     }).then(res => res.json())
 }
 
-export const updatePrompt = async (key: string, prompt: Partial<Prompt>) => {
+export const updatePrompt = async (
+    key: string,
+    prompt: Partial<Prompt>
+): Promise<{
+    success: boolean
+    key: string
+    prompt: Prompt | null
+    message: string
+}> => {
+    if (!await auth()) return { success: false, key, prompt: null, message: unauthorized }
+
     return fetch(`${process.env.API_URL}/api/prompts/${key}`, {
         method: "POST",
         headers: commonHeaders,
@@ -25,6 +42,7 @@ export const updatePrompt = async (key: string, prompt: Partial<Prompt>) => {
 }
 
 export const deletePrompt = async (promptName: string) => {
+    if (!await auth()) return { success: false, message: unauthorized }
     return fetch(`${process.env.API_URL}/api/prompts/${promptName}`, {
         method: "DELETE",
         headers: commonHeaders
@@ -32,6 +50,7 @@ export const deletePrompt = async (promptName: string) => {
 }
 
 export const importPrompts = async (prompts: Prompt[]) => {
+    if (!await auth()) return { message: unauthorized }
     return fetch(`${process.env.API_URL}/api/prompts`, {
         method: "PUT",
         headers: commonHeaders,
